tests: enable back buy ens name test
