+++ test/e2e/tests/transactions_tests/__init__.py
@@ -1,0 +1,3 @@
+import pytest
+
+marks = pytest.mark.transaction
